Support FIFO SQS queues via .withMessageGroupId()
To write messages into a FIFO queue from Spark, we should set the
MessageGroupId as a parameter of the SendMessageRequest.

I naively tried setting MessageGroupId with the existing msg_attributes
parameter, but it doesn't work with SQS. We cannot set the
`MessageGroupId` using `withMessageAttributes`. The `MessageGroupId` is
a specific parameter used by Amazon SQS to group and order messages in
a FIFO queue [docs.aws.amazon.com][0]. It is not the same as a message
attribute, which is custom metadata that you can attach to a message.

[0]: https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/using-messagegroupid-property.html